server: rename sql_backend_server -> sql_tcp_backend_server
